crawler/requirements.txt
httpx>=0.27
aiosqlite>=0.20
feedparser>=6.0
beautifulsoup4>=4.12
apscheduler>=3.10
plyer>=2.1
python-dotenv>=1.0
